Fixed almost everything

+++ Igrac.cpp
@@ -1,11 +1,11 @@
 #include "Igrac.h"
 
-void Igrac::povecajVrednost(int proc)
+void Igrac::povecajVrednost(float proc)
 {
 	vrednost += (proc / 100 * vrednost);
 }
 
-void Igrac::smanjiVrednost(int proc)
+void Igrac::smanjiVrednost(float proc)
 {
 	vrednost -= (proc / 100 * vrednost);
 }
